src/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
azure-common==1.1.28
azure-core==1.30.1
azure-search-documents==11.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
fastapi==0.110.2
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isodate==0.6.1
openai==1.23.6
packaging==24.0
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
requests==2.31.0
setuptools==68.2.2
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
wheel==0.41.2
